Move utility functions to AlphaWalletCore pod
