+++ AlphaWallet/Core/TokensAutodetector/TokensAutodetector.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import AlphaWalletCore
 import PromiseKit
 
 extension SingleChainTokensAutodetector {
@@ -77,7 +78,7 @@ class SingleChainTokensAutodetector: NSObject, TokensAutodetector {
             self?.autoDetectTransactedTokens()
             self?.autoDetectPartnerTokens()
         }
-    } 
+    }
 
     ///Implementation: We refresh once only, after all the auto detected tokens' data have been pulled because each refresh pulls every tokens' (including those that already exist before the this auto detection) price as well as balance, placing heavy and redundant load on the device. After a timeout, we refresh once just in case it took too long, so user at least gets the chance to see some auto detected tokens
     private func autoDetectTransactedTokens() {
@@ -237,7 +238,7 @@ extension SingleChainTokensAutodetector: AutoDetectTransactedTokensOperationDele
         return when(resolved: [fetchErc20Tokens, fetchNonErc20Tokens])
             .map(on: queue, { results in
                 return results.compactMap { $0.optionalValue }.flatMap { $0 }
-            }) 
+            })
     }
 }
