+++ modules/AlphaWalletCore/AlphaWalletCore.podspec
@@ -25,4 +25,5 @@ Pod::Spec.new do |s|
 
   s.frameworks       = 'Foundation'
 
+  s.dependency 'PromiseKit'
 end
