+++ modules/AlphaWalletCore/AlphaWalletCore/Extensions/PromiseKit+Extensions.swift
@@ -1,0 +1,20 @@
+//
+//  PromiseKit+Extensions.swift
+//  AlphaWalletOpenSea
+//
+//  Created by Hwee-Boon Yar on Apr/30/22.
+//
+
+import Foundation
+import PromiseKit
+
+extension PromiseKit.Result {
+    public var optionalValue: T? {
+        switch self {
+        case .fulfilled(let value):
+            return value
+        case .rejected:
+            return nil
+        }
+    }
+}
